install_sync_gateway.py deletes recreates buckets by default
https://github.com/couchbaselabs/mobile-testkit/issues/656#issuecomment-243860295
