--- libraries/provision/ansible/playbooks/tasks/flush-server-buckets.yml
@@ -1,23 +1,0 @@
-# Delete buckets
-- debug: msg={{ couchbase_server_primary_node }}
-- debug: msg="Couchbase cluster RAM {{ couchbase_server_cluster_ram }}"
-- debug: msg="Couchbase bucket RAM {{ couchbase_server_bucket_ram }}"
-
-- name: COUCHBASE SERVER | Delete buckets
-  shell: "{{ couchbase_server_home_path }}/bin/couchbase-cli bucket-delete --cluster={{ couchbase_server_primary_node }}:{{ couchbase_server_admin_port }} --user={{ couchbase_server_admin }} --password={{ couchbase_server_password }} --bucket={{ item }}"
-  register: bucket_flush
-  ignore_errors: true
-  with_items:
-    - data-bucket
-    - index-bucket
-  when: couchbase_server_node == couchbase_server_primary_node
-
-# Create buckets and wait
-- name: COUCHBASE SERVER | Create new buckets
-  shell: "{{ couchbase_server_home_path }}/bin/couchbase-cli bucket-create --cluster={{ couchbase_server_primary_node }}:{{ couchbase_server_admin_port }} --user={{ couchbase_server_admin }} --password={{ couchbase_server_password }} --bucket={{ item }} --bucket-type={{ couchbase_server_bucket_type }} --bucket-ramsize={{ couchbase_server_bucket_ram }} --bucket-replica={{ couchbase_server_bucket_replica }} --wait"
-  with_items:
-    - data-bucket
-    - index-bucket
-  when: couchbase_server_node == couchbase_server_primary_node
-
-
